requirements.txt
flask
pillow
numpy
requests
tensorflow
